WIP: define loan options
